requirements.txt
# Core
Django==1.3.1
psycopg2==2.4.1
PPyGIS
python-memcached
django-grappelli
django-filebrowser==3.3.0
#django-filebrowser-no-grappelli
django-tinymce
simplejson
PIL
South
docutils
django-extensions
django-static
slimmer
johnny-cache
django-model-utils
BeautifulSoup
markdown
FeedParser
-e git+git://github.com/pydanny/django-uni-form.git#egg=django-uni-form
#dutils
-e hg+https://bitbucket.org/salvator/django-admintools-bootstrap#egg=admintools_bootstrap

# Dev utils
django-debug-toolbar
django-sentry

# Misc
django-reversion
#pygraphviz # Enable this to generate ERD diagrams
django-mptt==0.4.2
django-tagging
django-tagging-autocomplete
django-smart-selects
django-haystack
whoosh
recaptcha-client
python-twitter
pygeoip

#CMS, blogging and other content management
django-flatblocks
django-cms
django-cms-search
cmsplugin-contact

# Social auth & account management
oauth2
python-openid
django-email-confirmation
# -e git+http://github.com/facebook/python-sdk.git#egg=facebook-sdk
-e git+https://github.com/pythonforfacebook/facebook-sdk.git#egg=facebook-sdk
#-e git+git://github.com/pennersr/django-allauth.git#egg=django-allauth
django-allauth

# Raw Jam forks of pluggable apps
-e git+git://github.com/rawjam/django-pagination.git#egg=django-pagination
